Simplify variables that user must set
Some of the variables in this script must match values that the
user provides to the SQL Server AlwaysOn Cluster template in the
Azure portal.

To make it easier to match these, relax our naming conventions
for those particular variables.

Also move these variables into a single section for easy reference.